Made the storage of components more flexible
Now the platform_file_path parameter of options.json points to a file
that must have a dict with one or more of these keys: platform, network,
node, processor. Each of these elements is a dict with named components.
The platform_name is the key of a component in the platform dict, and it
will be the root element for the simulation.

Components are read from the platform_file_path, if it is defined, or
from the library files. These are found in hdeeprm/data by default. Each
file is loaded into one of the previous dictionaries: platform, network,
node, processor. Each file contains a dictionary with components of the
same type; they can not be mixed, like in the platform_file_path.

The definition of a platform may combine elements from the
platform_file_path and from the library files.

The workload must be in json format. The SWF parser has been removed.